automcs150-backend/requirements.txt
acres==0.3.0
alembic==1.15.2
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
dnspython==2.7.0
email_validator==2.2.0
etelemetry==0.3.1
filelock==3.18.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
looseversion==1.3.0
lxml==5.3.2
Mako==1.3.9
MarkupSafe==3.0.2
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdfrw==0.4
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.28
pydot==3.0.4
PyJWT==2.10.1
PyMuPDF==1.25.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-safer==2.0
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
scipy==1.13.1
simplejson==3.20.1
six==1.17.0
SQLAlchemy==2.0.40
traits==7.0.2
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
